backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-telegram-bot
requests
gunicorn
